+++ src/gui/install-dialogs.ts
@@ -1,12 +1,16 @@
 import { FileCache } from '../cache'
 import { Lang } from '../lang-manager'
 import { LocalMods } from '../local-mods'
-import { InstallQueue, ModInstaller } from '../mod-installer'
-import { ModEntry, ModEntryLocal } from '../types'
+import { InstallQueue, ModInstaller, ModInstallerDownloadingProgress } from '../mod-installer'
+import { ModEntry, ModEntryLocal, ModEntryServer } from '../types'
 
 export function prepareModName(name: string) {
     return name.replace(/\\c\[\d]/g, '')
 }
+
+const white = '\\c[0]'
+const green = '\\c[2]'
+const yellow = '\\c[3]'
 
 export class ModInstallDialogs {
     static showModInstallDialog() {
@@ -19,7 +23,7 @@ export class ModInstallDialogs {
             return mods
                 .map(mod => {
                     const localVersion = LocalMods.getAllRecord()[mod.id]?.version
-                    return `- \\c[3]${prepareModName(mod.name)}\\c[0] ${localVersion ? `${localVersion} -> ` : ''}${mod.version}\n`
+                    return `- ${yellow}${prepareModName(mod.name)}${white} ${localVersion ? `${localVersion} -> ` : ''}${mod.version}\n`
                 })
                 .join('')
         }
@@ -28,11 +32,101 @@ export class ModInstallDialogs {
         const depsStr = deps.length > 0 ? `${Lang.dependencies}\n${modsToStr(deps)}` : ''
 
         const str = `${toInstallStr}${toUpdateStr}${depsStr}`
+        const modCount = toInstall.length + toUpdate.length + deps.length
 
-        let notAgain = false
-        const installModsFunc = () => {
-            if (notAgain) return
-            notAgain = true
+        const dialog = new modmanager.gui.MultiPageButtonBoxGui(undefined, undefined, [
+            {
+                name: Lang.install,
+                onPress: installModsFunc,
+            },
+            {
+                name: ig.lang.get('sc.gui.shop.cancel'),
+                onPress() {
+                    dialog.closeMenu()
+                },
+            },
+        ])
+        dialog.setContent(Lang.installButton.replace(/\[modCount\]/, modCount.toString()), [
+            {
+                content: [str],
+            },
+        ])
+
+        const STATUS = {
+            preparing: 0,
+            downloading: 1,
+            installing: 2,
+            done: 3,
+        } as const
+        const log: {
+            mod: ModEntryServer
+            progressFunc?: () => ModInstallerDownloadingProgress
+            status: (typeof STATUS)[keyof typeof STATUS]
+        }[] = []
+
+        const KiB = 1024
+        const MiB = KiB * 1024
+
+        function displayLog() {
+            const str = log
+                .sort((a, b) => a.status - b.status)
+                .map(({ mod, status, progressFunc }) => {
+                    let statusStr: string
+                    if (status == STATUS.preparing) {
+                        statusStr = 'Preparing...'
+                    } else if (status == STATUS.downloading) {
+                        const progress = progressFunc!()
+                        const [scale, unit] = progress.length >= MiB ? [MiB, 'MiB'] : [KiB, 'KiB']
+                        statusStr = `Downloading... ${yellow}${(progress.received / scale).round(0)}${white}/${yellow}${(progress.length / scale).round(0)}${white} ${unit}`
+                    } else if (status == STATUS.installing) {
+                        statusStr = 'Installing...'
+                    } else {
+                        statusStr = `${green}Done${white}`
+                    }
+
+                    return `${yellow}${mod.id}${white} ${mod.version} - ${statusStr}`
+                })
+                .join('\n')
+
+            dialog.setPageText(0, [str])
+            dialog.refreshPage()
+        }
+
+        function installModsFunc(): void {
+            const [installButton, cancelButton] = dialog.userButtons!
+            installButton.setActive(false)
+            cancelButton.setActive(false)
+
+            dialog.setPageHeader(0, Lang.installingModsHeader.replace(/\[modCount\]/, modCount.toString()))
+
+            const eventIndex = ModInstaller.eventListeners.push({
+                preparing(mod) {
+                    log.push({ mod, status: STATUS.preparing })
+                    displayLog()
+                },
+                downloading(mod, progressFunc) {
+                    const entry = log.find(e => e.mod == mod)!
+                    entry.status = STATUS.downloading
+                    entry.progressFunc = progressFunc
+                    displayLog()
+                },
+                installing(mod) {
+                    const entry = log.find(e => e.mod == mod)!
+                    entry.status = STATUS.installing
+                    displayLog()
+                },
+                done(mod) {
+                    const entry = log.find(e => e.mod == mod)!
+                    entry.status = STATUS.done
+                    displayLog()
+                },
+            })
+            const dialogUpdate = dialog.update.bind(dialog)
+            let frame = 0
+            dialog.update = function (this: typeof dialog) {
+                if (frame++ % 3 == 0) displayLog()
+                dialogUpdate()
+            }
 
             const toInstall = InstallQueue.values()
             ModInstaller.install(toInstall)
@@ -41,6 +135,7 @@ export class ModInstallDialogs {
                     modmanager.gui.menu &&
                         sc.Model.notifyObserver(modmanager.gui.menu, modmanager.gui.MENU_MESSAGES.UPDATE_ENTRIES)
 
+                    ModInstaller.eventListeners.splice(eventIndex, 1)
                     dialog.closeMenu()
 
                     sc.BUTTON_SOUND.shop_cash.play()
@@ -62,27 +157,6 @@ export class ModInstallDialogs {
                     })
                 })
         }
-
-        const dialog = new modmanager.gui.MultiPageButtonBoxGui(undefined, undefined, [
-            {
-                name: Lang.install,
-                onPress: installModsFunc,
-            },
-            {
-                name: ig.lang.get('sc.gui.shop.cancel'),
-                onPress() {
-                    dialog.closeMenu()
-                },
-            },
-        ])
-        dialog.setContent(
-            Lang.installButton.replace(/\[modCount\]/, (toInstall.length + toUpdate.length + deps.length).toString()),
-            [
-                {
-                    content: [str],
-                },
-            ]
-        )
         dialog.openMenu()
     }
 
@@ -121,7 +195,7 @@ export class ModInstallDialogs {
         if (deps.length > 0) {
             sc.Dialogs.showErrorDialog(
                 Lang.cannotUninstall.replace(/\[modName\]/, prepareModName(localMod.name)) +
-                    deps.map(mod => `- \\c[3]${prepareModName(mod.name)}\\c[0]\n`).join('')
+                    deps.map(mod => `- ${yellow}${prepareModName(mod.name)}${white}\n`).join('')
             )
             return false
         }
@@ -166,7 +240,7 @@ export class ModInstallDialogs {
         if (deps.length == 0) return true
         sc.Dialogs.showErrorDialog(
             Lang.cannotDisable.replace(/\[modName\]/, prepareModName(mod.name)) +
-                deps.map(mod => `- \\c[3]${prepareModName(mod.name)}\\c[0]\n`).join('')
+                deps.map(mod => `- ${yellow}${prepareModName(mod.name)}${white}\n`).join('')
         )
         return false
     }
@@ -179,7 +253,7 @@ export class ModInstallDialogs {
             sc.Dialogs.showYesNoDialog(
                 Lang.doYouWantToEnable
                     .replace(/\[modName\]/, prepareModName(mod.name))
-                    .replace(/\[mods\]/, deps.map(mod => `- \\c[3]${prepareModName(mod.name)}\\c[0]\n`).join('')),
+                    .replace(/\[mods\]/, deps.map(mod => `- ${yellow}${prepareModName(mod.name)}${white}\n`).join('')),
                 sc.DIALOG_INFO_ICON.QUESTION,
                 button => {
                     if (button.data == 0) {